refactor: 상품 select 로직 커스텀훅으로 분리
