+++ src/constants/products.ts
@@ -1,1 +1,6 @@
 export const CATEGORY = ['전체', '식료품', '패션잡화'];
+
+export const SORT: { [key: string]: string } = {
+  '낮은 가격 순': 'asc',
+  '높은 가격 순': 'desc',
+};
